+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 3.0 - Inside Joke (2019-06-13)
+- Private command framework added
+- Removed restrictions for Safebooru searches
+- Move changelog link back to GitHub
+- Update to bot list stats function
+- Remove timestamp from Booru results
+- Package updates
+
 ## 2.2 - Sauce Pls (2019-03-28)
 - [sagiri](https://www.npmjs.com/package/sagiri) package added
 - `m-source` command added